- Update Laravel Version 8. - Modificado composer.json. - Modificado Handler.
