+++ composer.json
@@ -8,12 +8,12 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.0.0",
+        "php": "^7.3.0",
         "barryvdh/laravel-dompdf": "^0.8.6",
         "fideloper/proxy": "^4.0",
-        "laravel/framework": "^6.2",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^1.2",
+        "laravel/ui": "^3.0",
         "laravelcollective/html": "^6.1",
         "luecano/numero-a-letras": "^2.0",
         "maatwebsite/excel": "^3.1",
@@ -21,11 +21,10 @@
         "spatie/laravel-permission": "^3.16"
     },
     "require-dev": {
-        "facade/ignition": "^1.4",
-        "fzaninotto/faker": "^1.4",
+        "facade/ignition": "^2.3.6",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^3.0",
-        "phpunit/phpunit": "^8.0"
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -39,12 +38,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\":"database/factories/",
+            "Database\\Seeders\\":"database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
