+++ database/seeders/ConceptoSeeder.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Concepto;
 use Illuminate\Database\Seeder;
