+++ database/seeders/ConfiguracionCategoriaSeeder.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Categoria;
 use App\ConfiguracionCategoria;
 use App\Empleado;
