--- database/seeds/CategoriaSeeder.php
@@ -1,21 +1,0 @@
-<?php
-
-use App\Categoria;
-use Illuminate\Database\Seeder;
-
-class CategoriaSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        Categoria::create([
-            'descripcion'=>'Administrador',
-            'salario_basico' => '60000',
-            'estado' => 'activo'
-        ]);
-    }
-}
